dom/tests/mochitest/mouselock/test_limitlessScroll.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=633602
Written by: Hyungryul Chun (hchun) - Student of DPS909 Fall 2011
-->
<head>
    <title>Test for Bug 633602: Scrolling limitlessly</title>
    <script type="application/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
    <script type="text/javascript" src="/tests/SimpleTest/EventUtils.js"></script>
    <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
</head>
<body>
    <a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=633602">Mozilla Bug 633602</a>
	
<div id="content">
</div>
<canvas id="canvas" width="150" height="150"></canvas>

    <pre id="test">
        <script type="application/javascript">
			
            /** Test for Bug 633602 **/
            SimpleTest.waitForExplicitFinish()
            SimpleTest.waitForFocus(function() {
				var xMove = screen.width;
				var yMove = screen.height;
                var pointer = navigator.pointer;
				var canvas = document.getElementById("canvas");
				var nConditions = 0;
				var xtestOK = false;
				var ytestOK = false;
				var movedX = 0;
				var movedY = 0;
				var startCounting = 0;
				SpecialPowers.setBoolPref("full-screen-api.allow-trusted-requests-only", false);
				
				document.addEventListener("mousemove", function (e) {
					if (startCounting === 1)
					{
						movedX = movedX + e.movementX;
						movedY = movedY + e.movementY;
					}					
				}, false);
				
				document.addEventListener("mozfullscreenchange", function (e) {     
					if (document.mozFullScreen && document.mozFullScreenElement === canvas) {
						pointer.lock(canvas);
						// Mouse should be able to move out of screen if in fullscreen mode
						// Move mouse's x axis off the screen (right edge)
						synthesizeMouseAtCenter(canvas, { type: "mousemove" }, window); 
						startCounting = 1;
						synthesizeMouse(canvas, xMove-1, yMove/2, { type: "mousemove" }, window); 
						synthesizeMouse(canvas, xMove-1, yMove/2, { type: "mousemove" }, window); 
						synthesizeMouse(canvas, xMove-1, yMove/2, { type: "mousemove" }, window); 
						
						// Move mouse's y axis off the screen (bottom edge)
						synthesizeMouse(canvas, xMove/2, yMove-1, { type: "mousemove" }, window); 
						synthesizeMouse(canvas, xMove/2, yMove-1, { type: "mousemove" }, window); 
						synthesizeMouse(canvas, xMove/2, yMove-1, { type: "mousemove" }, window);
						
						//is the delta x movement greater than the screen width?
						if (movedX > xMove){
							xtestOK = true;
						}	
						
						//is the delta y movement greater than the screen height?
						if (movedY > yMove){
							ytestOK = true;
						}

						
						ok(xtestOK, "The mouse should move beyond screen width!");
						is(movedX, (xMove-1-xMove/2)*3, "The mouse should move beyond the right edge of screen!");
						
						ok(ytestOK, "The mouse should move beyond screen heigt!")
						is(movedY, (yMove-1-yMove/2)*3, "The mouse should move beyond the bottom edge of screen!");
						
						movedX=0;
						movedY=0;
						// Move mouse's x axis off the screen (left edge)
						synthesizeMouse(canvas, 0, yMove/2, { type: "mousemove" }, window); 
						synthesizeMouse(canvas, 0, yMove/2, { type: "mousemove" }, window); 
						synthesizeMouse(canvas, 0, yMove/2, { type: "mousemove" }, window); 
						// Move mouse's y axis off the screen (top edge)
						synthesizeMouse(canvas, xMove/2, 0, { type: "mousemove" }, window); 
						synthesizeMouse(canvas, xMove/2, 0, { type: "mousemove" }, window); 
						synthesizeMouse(canvas, xMove/2, 0, { type: "mousemove" }, window);
						ok(xtestOK, "The mouse should move beyond screen width!")
						is(movedX, -1*((xMove/2)*3), "The mouse should move beyond the left edge of screen!");
						ok(ytestOK, "The mouse should move beyond screen heigt!")
						is(movedX, -1*((xMove/2)*3), "The mouse should move beyond the top edge of screen!");
						
						nConditions = 1;
						document.mozCancelFullScreen();
					} 
					else {
						if (nConditions) {
							// Finish test
							SimpleTest.finish();
						}
					}
				}, false);
				
				canvas.mozRequestFullScreen();
				
            });
        </script>
    </pre> 
</body>
</html>

	
